Create the device tree module
Add a new module that provides support for viewing and modifying
the device tree of Storage module. This module will provide full
functionality that is required by UI. Parts of this module will
be used in the partitioning module.
