+++ pyanaconda/modules/storage/devicetree/handler.py
@@ -1,5 +1,5 @@
 #
-# DBus interface for the device tree module
+# Handler of the device tree
 #
 # Copyright (C) 2019 Red Hat, Inc.
 #
@@ -17,14 +17,23 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
-from pyanaconda.dbus.interface import dbus_class
-from pyanaconda.modules.storage.devicetree.handler_interface import DeviceTreeHandlerInterface
-from pyanaconda.modules.storage.devicetree.viewer_interface import DeviceTreeViewerInterface
+from abc import abstractmethod, ABC
 
-__all__ = ["DeviceTreeInterface"]
+from pyanaconda.anaconda_loggers import get_module_logger
+
+log = get_module_logger(__name__)
+
+__all__ = ["DeviceTreeHandler"]
 
 
-@dbus_class
-class DeviceTreeInterface(DeviceTreeViewerInterface, DeviceTreeHandlerInterface):
-    """DBus interface for the device tree module."""
-    pass
+class DeviceTreeHandler(ABC):
+    """The viewer of the device tree."""
+
+    @property
+    @abstractmethod
+    def storage(self):
+        """The storage model.
+
+        :return: an instance of Blivet
+        """
+        return None
