+++ pyanaconda/modules/storage/devicetree/handler_interface.py
@@ -1,5 +1,5 @@
 #
-# DBus interface for the device tree module
+# DBus interface for the device tree handler
 #
 # Copyright (C) 2019 Red Hat, Inc.
 #
@@ -17,14 +17,14 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
-from pyanaconda.dbus.interface import dbus_class
-from pyanaconda.modules.storage.devicetree.handler_interface import DeviceTreeHandlerInterface
-from pyanaconda.modules.storage.devicetree.viewer_interface import DeviceTreeViewerInterface
+from pyanaconda.dbus.interface import dbus_interface
+from pyanaconda.dbus.template import InterfaceTemplate
+from pyanaconda.dbus.typing import *  # pylint: disable=wildcard-import
+from pyanaconda.modules.common.constants.interfaces import DEVICE_TREE_HANDLER
 
-__all__ = ["DeviceTreeInterface"]
+__all__ = ["DeviceTreeHandlerInterface"]
 
 
-@dbus_class
-class DeviceTreeInterface(DeviceTreeViewerInterface, DeviceTreeHandlerInterface):
-    """DBus interface for the device tree module."""
-    pass
+@dbus_interface(DEVICE_TREE_HANDLER.interface_name)
+class DeviceTreeHandlerInterface(InterfaceTemplate):
+    """DBus interface for the device tree handler."""
