+++ pyanaconda/modules/storage/storage_interface.py
@@ -21,11 +21,10 @@ from pyanaconda.modules.common.constants.services import STORAGE
 from pyanaconda.modules.common.base import KickstartModuleInterface
 from pyanaconda.dbus.interface import dbus_interface
 from pyanaconda.dbus.typing import *  # pylint: disable=wildcard-import
-from pyanaconda.modules.storage.devicetree import DeviceTreeInterface
 
 
 @dbus_interface(STORAGE.interface_name)
-class StorageInterface(KickstartModuleInterface, DeviceTreeInterface):
+class StorageInterface(KickstartModuleInterface):
     """DBus interface for Storage module."""
 
     def ResetWithTask(self) -> ObjPath:
